Raspberry Pi Camera module Projects
A series of scripts with some borrowed/found code that will be the basis
for setting up a raspberry Pi camera that will react when it detects a
face. TO DO, add a txt file which would have PCA(or other coordinates)
that can be recognized as a 'specific' person.